requirements.txt
appdirs==1.4.3
attrs==20.3.0
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
decorator==4.4.2
distlib==0.3.0
distro==1.4.0
future==0.18.2
html5lib==1.0.1
huepy==1.2.1
idna==2.8
imageio==2.9.0
imageio-ffmpeg==0.4.3
iniconfig==1.1.1
instabot==0.117.0
ipaddr==2.2.0
lockfile==0.12.2
mock==4.0.3
moviepy==1.0.3
msgpack==0.6.2
numpy==1.20.1
packaging==20.3
pep517==0.8.2
Pillow==8.1.1
pluggy==0.13.1
proglog==0.1.9
progress==1.5
py==1.10.0
pyparsing==2.4.6
PySocks==1.7.1
pytest==6.2.2
python-dotenv==0.15.0
pytoml==0.1.21
pytz==2021.1
requests==2.22.0
requests-toolbelt==0.9.1
responses==0.12.1
retrying==1.3.3
schedule==1.0.0
six==1.14.0
toml==0.10.2
tqdm==4.58.0
urllib3==1.25.8
webencodings==0.5.1
